backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.0.0
Django==4.2.20
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.13.2
